Added 3D spinning title text
The program can now load and display 3D models; the properties of the model (such as size, scale etc) can now be customised.

A 'placeholder' mesh for the title has been now added in.

The folder 'models' is now designed to hold any meshes to be used in the program.